[cses - 1661] Fix overflow.

+++ cses/1628.cc
@@ -4,7 +4,7 @@
 using namespace std;
 using ll = long long;
 using vi = vector<ll>;
-using mii = map<ll, int>;
+using mii = map<ll, ll>;
 
 ll x, s;
